requirements.txt
python-dotenv
fastapi
sentence-transformers
faiss-cpu
uvicorn
openai 